// src/app/shared/models/est_premi_basis_panen.model.ts
export class EstPremiBasisPanen {
	id?: number;

  blok_id?: number;
  tanggal_efektif?: string;

  bjr_dari?: number;
  bjr_sd?: number;
  basis_jjg?: number;
  basis_jjg_jumat?: number;

  lebih_basis1?: number;
  premi_lebih_basis1?: number;

  lebih_basis2?: number;
  premi_lebih_basis2?: number;

  lebih_basis3?: number;
  premi_lebih_basis3?: number;

  premi_brondolan?: number;
  hk_luas_panen?: number;
  premi_basis: number;
}